Update the PPU "swaps" logic to run from SQL
The big advantage here is that this is now driven by dm+d, rather than
trying to do clever things with the BNF code structure and a hand-rolled
list of substitutions. From a developer point of view the move from
Python code to SQL feels a bit retrograde. However there's a huge
advantage here in that the clinical informatics team can now manage this
logic themselves, including testing proposed changes in BigQuery.